+++ src/kernel_sdk/encoding/src/outbox.rs
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2022-2023 TriliTech <contact@trili.tech>
+// SPDX-FileCopyrightText: 2022-2024 TriliTech <contact@trili.tech>
 // SPDX-FileCopyrightText: 2023 Nomadic Labs <contact@nomadic-labs.com>
 //
 // SPDX-License-Identifier: MIT
@@ -69,6 +69,18 @@ impl<Expr: Michelson> From<Vec<OutboxMessageTransaction<Expr>>>
 {
     fn from(batch: Vec<OutboxMessageTransaction<Expr>>) -> Self {
         Self { batch }
+    }
+}
+
+impl<Expr: Michelson> From<OutboxMessageTransaction<Expr>> for OutboxMessage<Expr> {
+    fn from(transaction: OutboxMessageTransaction<Expr>) -> Self {
+        Self::AtomicTransactionBatch(vec![transaction].into())
+    }
+}
+
+impl<Expr: Michelson> From<Vec<OutboxMessageTransaction<Expr>>> for OutboxMessage<Expr> {
+    fn from(batch: Vec<OutboxMessageTransaction<Expr>>) -> Self {
+        Self::AtomicTransactionBatch(batch.into())
     }
 }