+++ src/kernel_sdk/CHANGES.md
@@ -5,6 +5,9 @@
 ### SDK
 - Add experimental support for compiling kernels to a Hermit RISC-V image behind the `proto-alpha` flag.
 - Add an experimental rollup host with an in-memory store behind the `experimental-host-in-memory-store` flag.
+- Add an `OutboxQueue` that can be used when more than 100 outbox messages are produced at a given level.
+- Add `From OutboxMessageTransaction`, `From OutboxMessageTransactionBatch` for `OutboxMessage` to simplify construction.
+- Fix the incomplete inbox on the first level of using `MockHost::default()`.
 
 ### Installer client/kernel